requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.6
cvzone==1.5.6
cycler==0.11.0
distlib==0.3.8
Django==5.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-shopping-cart==0.1
filelock==3.13.1
fonttools==4.38.0
gevent==22.10.2
greenlet==2.0.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
mysql-connector-python==8.0.31
mysqlclient==2.2.0
nikki-python==0.0.7
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==4.2.0
protobuf==3.20.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
razorpay==1.4.1
requests==2.28.2
Rx==3.2.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.25.0
websocket==0.2.1
websocket-client==1.5.1
zope.event==4.6
zope.interface==6.0
